+++ changelog.md
@@ -1,3 +1,15 @@
+## Version 0.23.02
+ - Hardened RClone /about path
+ - Hardened set/get config path
+ - Hardened Update system
+ - Hardened Invite system
+ - Fix issue with App Gate not checking 2FA consistently 
+ - Fix temperature reading when there's a warning (Thanks @BetterAndBetterII)
+ - Fix xram/swap shown under disks (Thanks @madejackson)
+
+## Version 0.23.01
+ - Allow client servers in Constellation
+
 ## Version 0.23.00
  - Remove Lungo, use SQLite instead
  - Remove MongoDB use SQLite / Postgres instead